fix: enforce integer validation for PSBT index in BIP122 schema
Updated the PSBT schema to ensure that the index field is an integer, enhancing data integrity and validation accuracy.

+++ src/data/methods/BIP122.methods.ts
@@ -32,7 +32,7 @@ export const bip122SignPsbtSchema = z.strictObject({
     .array(
       z.strictObject({
         address: z.string(),
-        index: z.number(),
+        index: z.number().int(),
         sighashTypes: z.array(z.number()).optional(),
       }),
     )
